docs: moved provider pagination docs snippets
